refactor: 게시물 단독 페이지 폴더 하나 더 생성
